# Use a 4-channel f32 pixel intermediate type for resizing.