This maintains precision and avoids a lot of unnecessary rounding to integers for intermediate values. This improves precision and performance.

On a 2950x:
Before: test imageops::sample::tests::bench_resize                     ... bench:   3,179,804 ns/iter (+/- 129,332) = 641 MB/s
After: test imageops::sample::tests::bench_resize                     ... bench:   1,563,759 ns/iter (+/- 87,143) = 1304 MB/s

Since there's no real generic [f32; 4] pixel type I just used Rgba32FImage as the intermediate type. The image is never converted to Rgba and the channel order is passed through, which is a bit weird but should be sound.

Fixes #1603